[tested] addDishButton, addDishToMenu, removeDish
addDishToMenu seems to work. Only adds 1 starter (or overwrites the
last), menu count updates as it should. Have not inspected menu array -
but should be correct. Also added "addToMenu" button that works - as
soon as we can read which id the radiobutton selected has.

Implemented and tested in this push:
addDishButton
removeDishButton
addDishToMenu
removeDishFromMenu